otherData refactor
