+++ src/data/otherData/carouselData.js
