+++ src/data/otherData/index.js
@@ -1,0 +1,3 @@
+export { locationsData } from './locationsData';
+export { homePageCards } from './homePageCards';
+export { carouselData } from './carouselData';
