+++ src/hooks/LocationCardModal.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import Button from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
 import Card from "react-bootstrap/Card";
-import { locationsData } from "../data/locationsData.js";
+import { locationsData } from "../data/otherData/locationsData.js";
 import Alert from "react-bootstrap/Alert";
 
 function LocationCardModal({ selectedLocation, showDelivery, showPickup }) {
